requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
